policial_federal/pyproject.toml
[project]
name = "agent-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
]
